backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
chromadb
PyPDF2
pdfplumber
sentence-transformers
tiktoken
numpy
langchain-text-splitters
